logistics-ai-backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
faiss-cpu
sentence-transformers
transformers
pydantic
python-docx
pypdf
langchain-huggingface
langchain-groq 
python-dotenv